projects: ad74414h: migrate from mbed to stm32 cmake
The mbed platform is EOL (issue #3186). Add the stm32 platform port
(SDP-K1 / sdp-ck1z: UART5 console, SPI5) with CMakeLists.txt, Kconfig,
the seven variant confs (basic, current_input_ext, current_input_loop,
current_output, digital_input_logic, digital_input_loop, digital_output),
board confs and the sdp-ck1z.ioc, and register the project in
projects/Kconfig. Remove the mbed platform files and the legacy Make
stack (builds.json, Makefile, src.mk). Document the CMake/no_os_build.py
flow in the README.
