+++ projects/Kconfig
@@ -66,6 +66,7 @@ source "projects/adf5611/Kconfig"
 source "projects/adf5902_sdz/Kconfig"
 source "projects/adf4377_sdz/Kconfig"
 source "projects/ad5460/Kconfig"
+source "projects/ad74414h/Kconfig"
 source "projects/ad7616-st/Kconfig"
 source "projects/ad7768-evb/Kconfig"
 source "projects/ad7768-1fmcz/Kconfig"
