--- projects/ad74414h/builds.json
@@ -1,25 +1,0 @@
-{
-  "mbed": {
-    "basic_example": {
-      "flags" : "EXAMPLE=basic"
-    },
-    "current_input_ext": {
-      "flags" : "EXAMPLE=current_input_ext"
-    },
-    "current_input_loop": {
-      "flags" : "EXAMPLE=current_input_loop"
-    },
-    "current_output": {
-      "flags" : "EXAMPLE=current_output"
-    },
-    "digital_input_logic": {
-      "flags" : "EXAMPLE=digital_input_logic"
-    },
-    "digital_input_loop": {
-      "flags" : "EXAMPLE=digital_input_loop"
-    },
-    "digital_output": {
-      "flags" : "EXAMPLE=digital_output"
-    }
-  }
-}
